refactor: consolidate middleware shared types into helpers
Move responseWriter and route pattern logic to helpers.go; wrapWriter
now guards against double-wrapping so chained middleware shares one
writer without scattered type assertions.
