+++ internal/middleware/helpers.go
@@ -1,0 +1,42 @@
+package middleware
+
+import "net/http"
+
+type responseWriter struct {
+	http.ResponseWriter
+	status int
+	wrote  bool
+}
+
+func (rw *responseWriter) WriteHeader(code int) {
+	if !rw.wrote {
+		rw.status = code
+		rw.wrote = true
+	}
+	rw.ResponseWriter.WriteHeader(code)
+}
+
+func (rw *responseWriter) Write(b []byte) (int, error) {
+	if !rw.wrote {
+		rw.WriteHeader(http.StatusOK)
+	}
+	return rw.ResponseWriter.Write(b)
+}
+
+func (rw *responseWriter) Unwrap() http.ResponseWriter {
+	return rw.ResponseWriter
+}
+
+func wrapWriter(w http.ResponseWriter) *responseWriter {
+	if rw, ok := w.(*responseWriter); ok {
+		return rw
+	}
+	return &responseWriter{ResponseWriter: w, status: http.StatusOK}
+}
+
+func routePattern(r *http.Request) string {
+	if r.Pattern != "" {
+		return r.Pattern
+	}
+	return "unknown"
+}
