+++ internal/middleware/logging.go
@@ -6,40 +6,17 @@ import (
 	"time"
 )
 
-type responseWriter struct {
-	http.ResponseWriter
-	status int
-	wrote  bool
-}
-
-func (rw *responseWriter) WriteHeader(code int) {
-	if !rw.wrote {
-		rw.status = code
-		rw.wrote = true
-	}
-	rw.ResponseWriter.WriteHeader(code)
-}
-
-func (rw *responseWriter) Write(b []byte) (int, error) {
-	if !rw.wrote {
-		rw.WriteHeader(http.StatusOK)
-	}
-	return rw.ResponseWriter.Write(b)
-}
-
-func (rw *responseWriter) Unwrap() http.ResponseWriter {
-	return rw.ResponseWriter
-}
-
 func Logging(logger *slog.Logger) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			start := time.Now()
-			rw := &responseWriter{ResponseWriter: w, status: http.StatusOK}
+			rw := wrapWriter(w)
 			next.ServeHTTP(rw, r)
+			route := routePattern(r)
 			logger.Info("request",
 				"method", r.Method,
 				"path", r.URL.Path,
+				"route", route,
 				"status", rw.status,
 				"duration_ms", time.Since(start).Milliseconds(),
 				"remote_addr", r.RemoteAddr,
